server, worker: Pass the target group information.
If the group information is removed, duplicate target names will
conflict and only the last one received will be used by the worker.

Sending the group information lets the worker know about everything. It
doesn't need to use the group names, but does allow it to receive
multiple targets with the same name (e.g. a ping and DNS check).